Move token name token to token package
