+++ src/us/kbase/auth2/service/ui/Tokens.java
@@ -34,7 +34,6 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 import us.kbase.auth2.lib.Authentication;
 import us.kbase.auth2.lib.Role;
-import us.kbase.auth2.lib.TokenName;
 import us.kbase.auth2.lib.exceptions.IllegalParameterException;
 import us.kbase.auth2.lib.exceptions.InvalidTokenException;
 import us.kbase.auth2.lib.exceptions.MissingParameterException;
@@ -43,6 +42,7 @@ import us.kbase.auth2.lib.exceptions.NoTokenProvidedException;
 import us.kbase.auth2.lib.exceptions.UnauthorizedException;
 import us.kbase.auth2.lib.storage.exceptions.AuthStorageException;
 import us.kbase.auth2.lib.token.IncomingToken;
+import us.kbase.auth2.lib.token.TokenName;
 import us.kbase.auth2.lib.token.TokenSet;
 import us.kbase.auth2.lib.token.TokenType;
 import us.kbase.auth2.lib.user.AuthUser;
